requirements.txt
streamlit
freesound-api
